stone: Add preliminary Reader
This is largely noop but adds a bunch of scaffolding to allow us to read from
a BufRead trait, dynamically loading the 32-byte header from an input stream
or returning an IO error.

We'll need to add our own results now as we need to define an error enum
for the instance in which we find invalid magic strings.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>