moss/registry/transaction: Add skeleton (pre-graph) code
We'll need these components in place in order to allow the correct
search logic for local sets vs installed vs repo-wide, and its a
fair chunk of change prior to bringing in the graph traversal
code.

In order to simplify things, we'll split this across multiple changes
for documentary purposes.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>