client/postblit: Reorganise to support system/transaction split
At a high level this is a simple change that allows system scope
triggers to run at a later stage than the transaction triggers.
These are run when `/usr` is activated.

For now, we still use a container when entering a non-`/` root to
ensure we don't fudge the host system, and we will add skip conditions
to triggers as and when we need them.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>