client/postblit: Rework trigger scopes to allow ephemeral roots
We rejigger the scope to incorporate the client scope and tightly
control the logic behind the bind paths, ensuring all possible paths
for trigger isolation and execution are properly accounted for.

As a result we can now take advantage of system + transaction triggers
from within ephemeral roots, powering the new boulder port.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>