triggers: Rework to ensure we compile early and order correctly
Unfortunately the old implementation was kinda buggy and the depsorting
eliminated further "copies" of a trigger from running. We now bucketize
per handler ID, giving implicit grouping which is then sorted using DFS
in our DAG.

After that its a case of mapping the IDs to the compiled match handlers,
having no further need for the paths or transient data.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>