repo: Generate release archive
This adds a script that generates a source archive containing vendored dependencies. This source archive is suitable for building in environments that do not have networking.

This archive is bit-for-bit reproducible and will always generate the exact same archive when ran on a given git commit with a given version of system tooling.

Signed-off-by: Reilly Brogan <reilly@reillybrogan.com>