boulder: Update cargo macros
Instead of building packages in debug profile (which requires workarounds for at least every package that uses rust-embed) build in release profile but with environmental variable overrides for the settings we need in order for RUSTFLAGS to take precedence.

This was tested by building several packages with different LTO profiles and verifying that they built as expected.

Also, add a few helper definitions for the artifact directory and the cargo options so that they can be used in other places.

Signed-off-by: Reilly Brogan <reilly@reillybrogan.com>