boulder: Add get-py-deps.py for python dependency analysis
This reads and sanitizes python dependencies from .dist-info or
.egg-info python installation directories.

It would be nice if this could be handled by native rust code but due to
the custom formatting of requires.txt in .egg-info directories as well
as the tricky handling of Require-Dist in .dist-info directories; makes
it quite challenging.

In the future once `uv` support parsing requires.txt correctly we could
replace this with those APIs.