handler/elf: Prevent soname clashes from private/rpath exports
This fixes issue #321 by confining the export/use of providers outside
of libdirs to a namespace, based on the knowledge they'll not live in
the toplevel FHS directories (usr/*).

This prevents a nasty issue whereby a package may ship its own vendored
version of say, libEGL.so, which will confuse the provider system when
at the global scope and cause removal of unrelated revdeps.

Additionally it was seen in Firefox in early testing and broke resolution
for build dependencies.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>