cli/info: Revamp output for missing details
We now use a nice formatted list for deps/providers, visually
breaking up some of the noise. We also munch through potentially
crappy descriptions to emit as an aligned paragraph in the CLI
output.

As part of issue #333 we unique the input set to not print the same
remote version that is installed locally, and add missing build and
source release fields.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>