boulder-data: Add fat-lto-objects flags and enable by default
These flags enable both traditional bitcode and LLVM/GCC intermediate object language to be built at the same time. This is necessary when using a static library built with LTO on LLVM/GCC with the other. Also, from some testing this also seems to improve the ability for LTO to minimize the binary size further.

This _does_ increase the size of the static archive files but not to a massive degree and it allows for more low-level LTO with packages that use the same toolchain and use static archives built in another package.

Tested with both LLVM and GCC, full LTO, thin LTO, and no LTO. Mold was also tested with both LLVM and GCC.

Signed-off-by: Reilly Brogan <reilly@reillybrogan.com>