boulder-data: Support mold linker
Mold is often noticeably faster than LLD (and much faster than BFD) and is already used by numerous packages for that reason. Make it more ergonomic to use by adding a top-level `mold` key which adds it as a builddep and sets environmental flags appropriately to use it.

Tested with both the gnu and LLVM toolchains.

Signed-off-by: Reilly Brogan <reilly@reillybrogan.com>