moss/client: Allow skipping triggers for moss state activate
We need to swap the rootfs pointer to the fstx before we actually
enter the system, so pivot_root isn't available to us. This makes
it very difficult to run any of the system triggers at this point.

What we can do is augment the moss CLI to expose a command to
manually re-run system triggers, and then use this slightly later
in the boot sequence.

For now, lets just get triggers skipped, as its sufficient to get
the new fstx booted.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>