boot: Enable automatic rollback boot menus
We now emit kernels found in each of the last 5 transactions to permit
the rollback mechanism. This required updated blsforme, and we still
special case the handling of the current transaction. This will change
in subsequent commits to ensure any boot operations that may fail will
happen against the *staging* tree before we ever activate, and defer the
deletions until that's worked.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>