boulder: Fix pattern matching
Boulder had a needless check for `starts_with` which was causing erroneous package splitting in several cases. For instance, consider the following simplified patterns:

```
- llvm-libs:
  paths:
    - /usr/lib/libLLVM.so.*

- llvm-devel:
  paths:
    - /usr/lib/libLLVM.so
```

With this setup the file `/usr/lib/libLLVM.so.18.1` should be patterned into `llvm-libs`, however since the `starts_with` match was checked first it would end up in `llvm-devel`. The point of this was to ensure that we could have "globs" that matched directories that matched subfiles/subdirectories and in order to preserve that behavior we instead match twice, once on the submitted pattern and again with `/**` appended to get a recursive glob.

Also while we're at it apply a minor adjustment to escape the directory so that all glob patterns work correctly.

Signed-off-by: Reilly Brogan <reilly@reillybrogan.com>