boulder/draft: Handle zip vs tar origins for `boulder new`
Right now we just specialise based on `.zip` suffix but we can
easily add special casing in time and perhaps even check the
actual mimetype. The idea is to reuse this approach for boulders
own unpack code and build a chain of extensions for the `.tar` case
to handle the archive mode and compression helper.

Right now, this unblocks zips for boulder new so yay.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>