boulder/macros: Allow actually removing the new flags
Option groups shouldn't really be using `disabled` as they can't be
excluded from the flag set, which is in turn bricking the nvidia
module build.

Additionally, the build-id stuff is mandated at the toolchain level.
For now the option is kept here as a toggleable but it doesn't really
make sense to be here at all. In future it could perhaps become an
option group to decide the build-id hash variety.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>